Editor: Unify the preview dropdown between post and site editors
