--- packages/block-editor/src/components/preview-options/style.scss
@@ -1,64 +1,0 @@
-.block-editor-post-preview__dropdown {
-	padding: 0;
-}
-
-.block-editor-post-preview__button-resize.block-editor-post-preview__button-resize {
-	padding-left: $button-size-small + $grid-unit-10 + $grid-unit-10;
-
-	&.has-icon {
-		padding-left: $grid-unit-10;
-	}
-}
-
-.block-editor-post-preview__dropdown-content {
-	&.edit-post-post-preview-dropdown {
-		.components-menu-group {
-			&:first-child {
-				padding-bottom: $grid-unit-10;
-			}
-			&:last-child {
-				margin-bottom: 0;
-			}
-		}
-	}
-
-	.components-menu-group + .components-menu-group {
-		padding: $grid-unit-10;
-	}
-}
-
-.edit-post-header__settings,
-.edit-site-header-edit-mode__actions {
-	@include break-small () {
-		.editor-post-preview {
-			display: none;
-		}
-	}
-}
-
-// Reduced UI.
-.edit-post-header.has-reduced-ui {
-	@include break-small() {
-		// Apply transition to first two buttons.
-		.edit-post-header__settings .editor-post-save-draft,
-		.edit-post-header__settings .editor-post-saved-state,
-		.edit-post-header__settings .block-editor-post-preview__button-toggle {
-			transition: opacity 0.1s linear;
-			@include reduce-motion("transition");
-		}
-
-		// Zero out opacity unless hovered.
-		&:not(:hover) {
-			.edit-post-header__settings .editor-post-save-draft,
-			.edit-post-header__settings .editor-post-saved-state,
-			.edit-post-header__settings .block-editor-post-preview__button-toggle {
-				opacity: 0;
-			}
-
-			// ... or opened.
-			.edit-post-header__settings .block-editor-post-preview__button-toggle.is-opened {
-				opacity: 1;
-			}
-		}
-	}
-}
