+++ packages/edit-post/src/components/header/style.scss
@@ -294,7 +294,7 @@
 
 		& > .edit-post-header__toolbar .edit-post-header-toolbar__inserter-toggle,
 		& > .edit-post-header__toolbar .edit-post-header-toolbar__document-overview-toggle,
-		& > .edit-post-header__settings > .block-editor-post-preview__dropdown,
+		& > .edit-post-header__settings > .editor-preview-dropdown,
 		& > .edit-post-header__settings > .interface-pinned-items {
 			display: none;
 		}
