+++ packages/editor/src/components/preview-dropdown/style.scss
@@ -1,0 +1,5 @@
+.editor-preview-dropdown__button-external {
+	width: 100%;
+	display: flex;
+	justify-content: space-between;
+}
