+++ packages/editor/src/private-apis.js
@@ -7,6 +7,7 @@ import { lock } from './lock-unlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
 import useBlockEditorSettings from './components/provider/use-block-editor-settings';
 import PostPanelRow from './components/post-panel-row';
+import PreviewDropdown from './components/preview-dropdown';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -14,6 +15,7 @@ lock( privateApis, {
 	ExperimentalEditorProvider,
 	EntitiesSavedStatesExtensible,
 	PostPanelRow,
+	PreviewDropdown,
 
 	// This is a temporary private API while we're updating the site editor to use EditorProvider.
 	useBlockEditorSettings,
