+++ packages/editor/src/style.scss
@@ -23,5 +23,6 @@
 @import "./components/post-url/style.scss";
 @import "./components/post-visibility/style.scss";
 @import "./components/post-trash/style.scss";
+@import "./components/preview-dropdown/style.scss";
 @import "./components/table-of-contents/style.scss";
 @import "./components/template-validation-notice/style.scss";
